refactor: modularize the jwt decode, pre process s3 data and upload to s3
